Move freeze to internals
